equateModels(sbmlr)	R Documentation

Check the equality of the species and reactions of two SBML models

Description

This function tests the equivalence of two models with respect to the species and reaction data frames generated by summary.

Usage


equateModels(model1,model2)

Arguments

model1 	 The first of the two model objects of class SBML which are to be compared.

model2 	 The second model object.

Value

A list containing the following two boolean dataframes

species 	 The equality of species information tabularized as a data frame.

reactions 	 The equality of reaction information tabularized as a dataframe.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

summary.SBML

Examples


library(SBMLR)  
curto1=readSBMLR(file.path(system.file(package="SBMLR"), "models/curto.r"))  
curto2=readSBML(file.path(system.file(package="SBMLR"), "models/curto.xml"))  
equateModels(curto1,curto2)


[Package sbmlr version 1.31.0 Index]