readSBML(SBMLR)	R Documentation

Convert an SBML file into an R model object of class SBML

Description

This function converts an SBML level 2 file into a corresponding R model structure of class SBML.

Usage


readSBML(filename)

Arguments

filename 	 An SBML level 2 model input file.

Details

A limited subset of SBML level 2 models is currently supported, e.g. events and function definitions are not covered.

Value

A corresponding SBML model object in R.

Note

This function replaces read.SBML of older versions.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

readSBMLR

Examples


library(SBMLR)
library(odesolve)
curtoX=readSBML(file.path(system.file(package="SBMLR"), "models/curto.xml"))  
curtoR=readSBMLR(file.path(system.file(package="SBMLR"), "models/curto.r"))  
curtoX==curtoR

[Package SBMLR version 1.29.0 Index]