readSBMLR(SBMLR)	R Documentation

Convert an SBMLR file into an R model object of class SBML

Description

This function converts an SBMLR model definition in filename into a corresponding returned SBML model structure.

Usage


readSBMLR(filename)

Arguments

filename 	 An SBMLR model definition file.

Details

A limited subset of SBML level 2 models is currently supported, e.g. events and function definitions are not covered.

Value

A corresponding SBML model object in R.

Note

This function replaces the use of source in older versions of SBMLR. It includes rate law and rule string to function, expression and MathML mappings.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

readSBML

Examples


library(SBMLR)
library(odesolve)
curtoX=readSBML(file.path(.path.package("SBMLR"), "models/curto.xml"))  
curtoR=readSBMLR(file.path(.path.package("SBMLR"), "models/curto.r"))  
curtoX==curtoR

[Package SBMLR version 1.26.0 Index]