saveSBMLR(SBMLR)	R Documentation

Save an R model object of class SBML as an SBMLR file

Description

This function converts SBML model object in R into an SBMLR model definition file.

Usage


saveSBMLR(model,filename)

Arguments

model 	 The SBML model object to be mapped into the SBMLR model definition file.

filename 	 The file name of the destination SBMLR model definition file.

Value

No value returned.

Warning

SBML events and function definitions are NOT implemented.

Note

Similar to saveSBML, the file is written incrementally.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

References

Radivoyevitch, T. A two-way interface between limited Systems Biology Markup Language and R. BMC Bioinformatics 5, 190 (2004).

See Also

saveSBML

Examples


library(SBMLR)
library(odesolve)
curto=readSBMLR(file.path(.path.package("SBMLR"), "models/curto.r"))  
saveSBMLR(curto,file.path(.path.package("SBMLR"), "models/curtoR.r"))  
curtoR=readSBMLR(file.path(.path.package("SBMLR"), "models/curtoR.r"))  

curto==curtoR

[Package SBMLR version 1.26.0 Index]