summary.SBML(SBMLR)	R Documentation

Get summary data frames from an SBML model

Description

This function extracts species and reaction information from a model of class SBML and tabularizes them as two R data frames.

Usage


summary.SBML(object, ...)

Arguments

object 	 A model object of class SBML from which information is to be extracted.

... 	 For compatibility with summary of the base package.

Value

A list containing the following named dataframe elements

species 	 Species information (i.e. names, ICs, BCs, and compartments) as a data frame.

reactions 	 Reaction information tabularized as a dataframe, including string laws and initial fluxes.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

==.SBML

Examples


library(SBMLR)  
curto=readSBMLR(file.path(.path.package("SBMLR"), "models/Curto.r"))  
summary(curto)


[Package SBMLR version 1.21 Index]