==.SBML(SBMLR)	R Documentation

Check the equality of the summary data frames of two SBML models

Description

This function tests the equivalence of two models with respect to their summary data frames.

Usage


"==.SBML"(model1, model2, ...)

Arguments

model1 	 The first of the two model objects of class SBML which are to be compared.

model2 	 The second model object.

... 	 For compatibility with == of the base package.

Value

A list containing the following named dataframe elements as from summary, but now with boolean entries

species 	 The equality of species information tabularized as a data frame.

reactions 	 The equality of reaction information tabularized as a dataframe.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

summary.SBML

Examples


library(SBMLR)  
curto1=readSBMLR(file.path(.path.package("SBMLR"), "models/curto.r"))  
curto2=readSBML(file.path(.path.package("SBMLR"), "models/curto.xml"))  
curto1==curto2


[Package SBMLR version 1.26.0 Index]