saveSBML(SBMLR)	R Documentation

Saves an R model object as an SBML file

Description

This function converts a class SBML model object in R into an SBML level 2 file.

Usage


saveSBML(model,filename)

Arguments

model 	 The model object in R.

filename 	 The name of the SBML file

Details

The output file is SBML level 2.

Value

No value returned.

Warning

SBML events and function definitions are NOT implemented.

Note

For speed, the SBML file is written incrementally, rather than first built as a DOM in R and then saved using xmlSave.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

References

Radivoyevitch, T. A two-way interface between limited Systems Biology Markup Language and R. BMC Bioinformatics 5, 190 (2004).

See Also

saveSBMLR

Examples


library(SBMLR)
library(odesolve)
curtoR=readSBMLR(file.path(.path.package("SBMLR"), "models/curto.r"))  
saveSBML(curtoR,file.path(.path.package("SBMLR"), "models/curtoR.xml")) 
curtoX=readSBML(file.path(.path.package("SBMLR"), "models/curtoR.xml"))  

curtoX==curtoR

[Package SBMLR version 1.26.0 Index]