summary.SBML(SBMLR)	R Documentation

Get summary information from an SBML model

Description

This function extracts information from a model of class SBML and returns it as a list. The list includes species and reaction information tabularized as data frames.

Usage


summary.SBML(object, ...)

Arguments

object 	 A model object of class SBML from which information is to be extracted.

... 	 For compatibility with summary of the base package.

Value

A list containing the following elements

BC 	 A logical vector indicating which species are not state variables, i.e. which species are boundary conditions or auxillary variables.

y0 	 The initial state (boundary conditions excluded!).

nStates 	 The length of the state vector, i.e. the number of system states.

S0 	 The full set of species initial values.

nReactions 	 The number of reactions.

nSpecies 	 The number of species, including states, boundary conditions and possibly auxillary variables such as the total concentration of dihydofolate reductase in the morrison.r model.

incid 	 The incidence/stoichiometry matrix. This usually contains ones and minus ones corresponding to fluxes either synthesizing or degrading (respectively) a state variable chemical species. This matrix multiplied by the flux vector on its right yields the corresponding concentration state variable time derivatives.

species 	 Species information (i.e. names, ICs, BCs, and compartments) as a data frame.

reactions 	 Reaction information tabularized as a dataframe, including string laws and initial fluxes.

Note

The list output can be attached to immediately define many model variables of interest.

Author(s)

Tomas Radivoyevitch (radivot@hal.cwru.edu)

See Also

==.SBML

Examples


library(SBMLR)  
curto=readSBMLR(file.path(system.file(package="SBMLR"), "models/curto.r"))  
summary(curto)


[Package SBMLR version 1.29.0 Index]